+++ analysis/src/Shared.ml
@@ -5,18 +5,14 @@ let tryReadCmt cmt =
   else
     match Cmt_format.read_cmt cmt with
     | exception Cmi_format.Error err ->
-      Log.log
-        ("Failed to load " ^ cmt ^ " as a cmt w/ ocaml version " ^ "406"
-       ^ ", error: "
-        ^
-        (Cmi_format.report_error Format.str_formatter err;
-         Format.flush_str_formatter ()));
+      let error_message =
+        Cmi_format.report_error Format.str_formatter err;
+        Format.flush_str_formatter ()
+      in
+      Log.log ("Invalid cmt format " ^ cmt ^ ": " ^ error_message);
       None
     | exception err ->
-      Log.log
-        ("Invalid cmt format " ^ cmt
-       ^ " - probably wrong ocaml version, expected " ^ Config.version ^ " : "
-       ^ Printexc.to_string err);
+      Log.log ("Invalid cmt format " ^ cmt ^ ": " ^ Printexc.to_string err);
       None
     | x -> Some x