+++ compiler/ext/config.ml
@@ -1,5 +1,3 @@
-let version = "4.06.1+BS"
-
 (* This resolves the location of the standard library starting from the location of bsc.exe,
   handling different supported package layouts. *)
 let standard_library =
@@ -31,8 +29,6 @@ let standard_library =
   | _ :: _ :: _ :: _ :: rest -> build_path rest ["lib"; "ocaml"]
   | _ -> ""
 
-let standard_library_default = standard_library
-
 let cmi_magic_number = "Caml1999I022"
 
 and ast_impl_magic_number = "Caml1999M022"
@@ -42,20 +38,3 @@ and ast_intf_magic_number = "Caml1999N022"
 and cmt_magic_number = "Caml1999T022"
 
 let load_path = ref ([] : string list)
-
-(* This is normally the same as in obj.ml, but we have to define it
-   separately because it can differ when we're in the middle of a
-   bootstrapping phase. *)
-
-let print_config oc =
-  let p name valu = Printf.fprintf oc "%s: %s\n" name valu in
-  p "version" version;
-  p "standard_library_default" standard_library_default;
-  p "standard_library" standard_library;
-
-  (* print the magic number *)
-  p "cmi_magic_number" cmi_magic_number;
-  p "ast_impl_magic_number" ast_impl_magic_number;
-  p "ast_intf_magic_number" ast_intf_magic_number;
-  p "cmt_magic_number" cmt_magic_number;
-  flush oc