+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Remove unused `-bs-D` and `-bs-list-conditionals` flags. https://github.com/rescript-lang/rescript/pull/7631
 - Remove obsolete jsx options. https://github.com/rescript-lang/rescript/pull/7633
 - Remove obsolete option `-bs-unsafe-empty-array`. https://github.com/rescript-lang/rescript/pull/7635
+- Clean up `config.ml`. https://github.com/rescript-lang/rescript/pull/7636
 
 # 12.0.0-beta.1